% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posfunc.R
\name{posfunc}
\alias{posfunc}
\title{Probability of Success (Formula Method)}
\usage{
posfunc(hr_true = 0.75, hr_threshold = 0.85, n_events = 220)
}
\arguments{
\item{hr_true}{True hazard ratio which you assumed.}

\item{hr_threshold}{Hazard ratio threshold. It will be used to calculate Pr(hr < hr_threshold).}

\item{n_events}{PFS/OS events planned.}
}
\value{
Prints Pr(hr < hr_threshold) to console.
}
\description{
Calculating probability of success using formula method in oncology clinical trials.
}
\examples{
\dontrun{
posfunc(hr_true = 0.75, hr_threshold = 0.85, n_events = 220)
}
}
