% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pospack.R
\name{pospack}
\alias{pospack}
\title{Probability of Success (Package Method)}
\usage{
pospack(
  type = "surv",
  nsamples = 2,
  null.value = 1,
  alternative = "less",
  D = 197,
  a = 1,
  hr.prior = 0.667,
  D.prior = 9999999,
  succ.crit = "trial",
  alpha.final = 0.025
)
}
\arguments{
\item{type}{Type of the endpoint. "surv" for survival.}

\item{nsamples}{Number of samples.}

\item{null.value}{The specified value under null hypothesis.}

\item{alternative}{Direction of alternate hypothesis. "greater" or "less".}

\item{D}{Total number of events at final analysis.}

\item{a}{a = 1.}

\item{hr.prior}{Mean value of prior distribution for HR.}

\item{D.prior}{Prior D value.}

\item{succ.crit}{"trial" or "clinical".}

\item{alpha.final}{Alpha at final analysis.}
}
\value{
A list containing PoS results.
}
\description{
Calculating probability of success using LongCART method in oncology clinical trials.
}
\examples{
\dontrun{
pospack(type="surv", nsamples=2, null.value=1, alternative="less",
        D=197, a=1, hr.prior=0.667, D.prior=9999999,
        succ.crit="trial", alpha.final=0.025)
}
}
